requirements.txt
# Specify Python package requirements for your project here (e.g., Mako==1.1.1). If your project doesn't require these, you can leave this file unchanged or delete it.
selenium
undetected_chromedriver
python-dotenv
google-genai
PdfReader
pypdf